+++ src/libwikt/format3reader.h
@@ -40,15 +40,15 @@ public:
   /// @brief Returns content of an entry.
   /// @param offset
   ///   Offset to list of all entries. 0 <= offset < entryCount().
-  virtual QString source(int offset);
+  virtual QString source(quint32 offset);
   /// @param offset
   ///  Offset between 0 and file size.
-  virtual QString sourceDirect(int offset);
+  virtual QString sourceDirect(qint64 offset);
 
   /// Returns number of entries in the file.
-  int entryCount() const { return _entryCount; }
+  quint32 entryCount() const { return _entryCount; }
 
-  typedef QMap<QString, size_t> EntryMap;
+  typedef QMap<QString, qint64> EntryMap;
   const EntryMap &entries() const { return _links; }
 
 protected:
